+++ spacy/__init__.py
@@ -5,7 +5,7 @@ import sys
 # set library-specific custom warning handling before doing anything else
 from .errors import setup_default_warnings
 
-setup_default_warnings()
+setup_default_warnings()  # noqa: E402
 
 # These are imported as part of the API
 from thinc.api import prefer_gpu, require_gpu, require_cpu  # noqa: F401
