+++ spacy/tests/doc/test_creation.py
@@ -69,4 +69,4 @@ def test_create_with_heads_and_no_deps(vocab):
     words = "I like ginger".split()
     heads = list(range(len(words)))
     with pytest.raises(ValueError):
-        doc = Doc(vocab, words=words, heads=heads)
+        Doc(vocab, words=words, heads=heads)
