+++ spacy/tests/regression/test_issue5001-5500.py
@@ -69,9 +69,12 @@ def test_issue5082():
 
 
 def test_issue5137():
-    @Language.factory("my_component")
+    factory_name = "test_issue5137"
+    pipe_name = "my_component"
+
+    @Language.factory(factory_name)
     class MyComponent:
-        def __init__(self, nlp, name="my_component", categories="all_categories"):
+        def __init__(self, nlp, name=pipe_name, categories="all_categories"):
             self.nlp = nlp
             self.categories = categories
             self.name = name
@@ -86,13 +89,13 @@ def test_issue5137():
             pass
 
     nlp = English()
-    my_component = nlp.add_pipe("my_component")
+    my_component = nlp.add_pipe(factory_name, name=pipe_name)
     assert my_component.categories == "all_categories"
     with make_tempdir() as tmpdir:
         nlp.to_disk(tmpdir)
-        overrides = {"components": {"my_component": {"categories": "my_categories"}}}
+        overrides = {"components": {pipe_name: {"categories": "my_categories"}}}
         nlp2 = spacy.load(tmpdir, config=overrides)
-        assert nlp2.get_pipe("my_component").categories == "my_categories"
+        assert nlp2.get_pipe(pipe_name).categories == "my_categories"
 
 
 def test_issue5141(en_vocab):
